requirements.txt
flask
openai
azure-cognitiveservices-speech
azure-communication-calling